# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WasteWise API",
    "version": "1.0.0",
...
    "contact": {
      "name": "WasteWise Support",
      "email": "support@wastewise.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wastewise.com/terms"
    }
  },
  "servers": [
    {
      "url": "https://wastewise.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code"
          },
          "details": {
            "type": "object",
            "description": "Additional error context",
            "additionalProperties": true
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "description": "Items per page"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        }
      },
...
      "AnalysisJob": {
        "type": "object",
        "required": [
          "id",
          "status",
          "job_type",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Job ID"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "complete_analysis",
              "invoice_extraction",
              "contract_extraction",
              "regulatory_research"
            ],
            "description": "Type of analysis job"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ],
            "description": "Current job status"
          },
          "progress_percent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Completion percentage"
          },
          "current_step": {
            "type": "string",
            "description": "Current processing step",
            "example": "Extracting invoice data"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "description": "Processing start time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "Processing completion time"
          },
          "error_message": {
            "type": "string",
            "description": "Error message if job failed"
          },
          "error_code": {
            "type": "string",
            "description": "Error code if job failed"
          },
          "result_data": {
            "type": "object",
            "description": "Analysis results (available when status is completed)",
            "additionalProperties": true
          },
          "ai_usage": {
            "type": "object",
            "description": "AI token usage and cost tracking",
            "properties": {
              "input_tokens": {
                "type": "integer"
              },
              "output_tokens": {
                "type": "integer"
              },
              "total_cost": {
                "type": "number"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Authentication required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Insufficient permissions",
              "code": "FORBIDDEN"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Resource not found",
              "code": "NOT_FOUND"
            }
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-indexed)",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "Items per page",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "JobIdParam": {
        "name": "id",
        "in": "path",
        "description": "Job ID",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/jobs": {
      "get": {
        "summary": "List all jobs (admin)",
        "description": "Get paginated list of all analysis jobs across all users. Admin only.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
              ]
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AnalysisJob"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/admin/jobs/{id}": {
      "get": {
        "summary": "Get job details (admin)",
        "description": "Get detailed information about any job. Admin only.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/JobIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Job details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJob"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/admin/jobs/{id}/retry": {
      "post": {
        "summary": "Retry failed job (admin)",
        "description": "Create a new job with the same parameters as a failed job. Admin only.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/JobIdParam"
          }
        ],
        "responses": {
          "201": {
            "description": "Retry job created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "originalJobId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "example": "pending"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/admin/jobs/monitoring": {
      "get": {
        "summary": "Get job queue metrics (admin)",
        "description": "Real-time metrics about job queue health and performance. Admin only.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Queue metrics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "queue": {
                      "type": "object",
                      "properties": {
                        "pending": {
                          "type": "integer",
                          "description": "Jobs waiting to be processed"
                        },
                        "processing": {
                          "type": "integer",
                          "description": "Jobs currently being processed"
                        },
                        "completed_last_hour": {
                          "type": "integer"
                        },
                        "failed_last_hour": {
                          "type": "integer"
                        },
                        "avg_processing_time_seconds": {
                          "type": "number",
                          "description": "Average processing time"
                        }
                      }
                    },
                    "workers": {
                      "type": "object",
                      "properties": {
                        "active": {
                          "type": "integer",
                          "description": "Number of active workers"
                        },
                        "idle": {
                          "type": "integer",
                          "description": "Number of idle workers"
                        }
                      }
                    },
                    "performance": {
                      "type": "object",
                      "properties": {
                        "jobs_per_minute": {
                          "type": "number",
                          "description": "Current throughput"
                        },
                        "avg_wait_time_seconds": {
                          "type": "number",
                          "description": "Average time jobs spend waiting"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "queue": {
                    "pending": 5,
                    "processing": 2,
                    "completed_last_hour": 45,
                    "failed_last_hour": 1,
                    "avg_processing_time_seconds": 135.5
                  },
                  "workers": {
                    "active": 2,
                    "idle": 1
                  },
                  "performance": {
                    "jobs_per_minute": 0.5,
                    "avg_wait_time_seconds": 15.3
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
    "/api/admin/audit": {
      "get": {
        "summary": "Get audit log (admin)",
        "description": "Retrieve audit log of admin actions. Admin only.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "action",
            "in": "query",
            "description": "Filter by action type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Filter by user who performed action",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "user_id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "action": {
                                "type": "string"
                              },
                              "resource_type": {
                                "type": "string"
                              },
                              "resource_id": {
                                "type": "string"
                              },
                              "changes": {
                                "type": "object"
                              },
                              "timestamp": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    }
  }
}